+++ tests/integration/specification_model_test.py
@@ -17,7 +17,6 @@ from rflx.model import (
     Message,
     Model,
     ModularInteger,
-    Private,
     Session,
     State,
     Transition,
@@ -605,9 +604,8 @@ def test_consistency_specification_parsing_generation(tmp_path: Path) -> None:
         ],
         [
             decl.ChannelDeclaration("X", readable=True, writable=True),
-            decl.TypeDeclaration(Private("Test::T")),
-            decl.FunctionDeclaration("F", [], "Test::T"),
-            decl.FunctionDeclaration("G", [decl.Argument("P", "Test::T")], BOOLEAN.identifier),
+            decl.FunctionDeclaration("F", [], "Test::Tag"),
+            decl.FunctionDeclaration("G", [decl.Argument("P", "Test::Tag")], BOOLEAN.identifier),
         ],
         [BOOLEAN, OPAQUE, tag, length, message],
     )